refactor plugin registry and add docstrings
