+++ src/core/MSSCore.vcxproj.filters
@@ -113,6 +113,9 @@
     <ClInclude Include="include\pipeline\sinks\recording_defaults.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="include\utils\plugins_utils.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\controllers\maincontroller.cpp">
@@ -143,6 +146,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="src\pipeline\sessionpipeline.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="src\sdk\plugins\pluginloader.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
